requirements.txt
streamlit
sqlite3
datetime
typing
openai
python-dotenv